+++ package.json
@@ -23,7 +23,6 @@
 		"qt3tests": "ts-mocha --paths -p test/tsconfig.json test/qt3tests.ts",
 		"qt3testsxqueryx": "ts-mocha --paths -p test/tsconfig.json test/qt3testsXQueryX.ts",
 		"test": "ts-mocha --require test/testhook.js \"test/specs/**/*.ts\" --project test/tsconfig.json",
-		"jscodegentests": "ts-mocha --require test/testhook.js \"test/jsCodegen/**/*.ts\" --project test/tsconfig.json",
 		"xqutstests": "ts-mocha --paths -p test/tsconfig.json test/xqutsTests.ts",
 		"xqutstestsxqueryx": "ts-mocha --paths -p test/tsconfig.json test/xqutsTestsXQueryX.ts",
 		"browsertests": "karma start",
